Add validation and complete deployment
